Respect memberships in Grouper for unmanaged groups (CO-1531)

+++ app/Plugin/GrouperProvisioner/Model/CoGrouperProvisionerTarget.php
@@ -683,14 +683,17 @@ FROM
     }
 
     // Loop over the Grouper group memberships and delete the user
-    // from any groups not passed in as provisioning data.
+    // from any groups not passed in as provisioning data, but only
+    // if this is a group managed by this plugin.
     foreach($grouperGroups as $g) {
       if(!(in_array($g, $registryGroups))) {
-        try {
-          $grouper->deleteMember($g, $subject);
-        } catch (GrouperRestClientException $e) {
-          // Log the failure but go onto the next group.
-          $this->log("GrouperProvisioner unable to remove subject $subject from group $g");
+        if($this->CoGrouperProvisionerGroup->isManaged($g)) {
+          try {
+            $grouper->deleteMember($g, $subject);
+          } catch (GrouperRestClientException $e) {
+            // Log the failure but go onto the next group.
+            $this->log("GrouperProvisioner unable to remove subject $subject from group $g");
+          }
         }
       }
     }
